Create Installs_Strongswan_n_Quagga.sh
This shell script is a part of a project I am working on which automates IPSec site to site VPN creation with AWS. This script installs Strongswan and Quagga. It also makes the changes in .conf files of quagga and make is ready to deploy IPsec VPN with BGP routing.